Refactor project structure: separate core logic and multi-channel handlers (email, WhatsApp, Telegram)
